EA-8839: Update to version 1.1.24 and make libtdsodbc.so available
The provided RPMs have been setup to be organized in the same manner
as the freetds RPMs provided by EPEL.  The EPEL RPMs are setup as
follows:

    freetds - Contains binary executables, man pages, and some
              light documentation in the form of a user guide.

    freetds-libs - Contains libraries (including libtdsodbc.so),
                   some conf files, and some odbc samples.

    freetds-devel - Contains header files, libraries, & samples.

    freetds-doc - Contains the bulk of the documentation files.

We now have ea-freetds, ea-freetds-libs, ea-freetds-devel, and
ea-freetds-doc which all put the appropriate files under /opt/cpanel.

Previously, our ea-freetds also contained everything in ea-freetds-devel
and ea-freetds-doc.  The ea-freetds-devel RPM we were shipping was
completely redundant to the ea-freetds package.  And, we were not
building anything that is included with ea-freetds-lib.